Match links support custom routing.
Closes #31

When using a Progressive Web Application router such as React Router,
the application code needs access to both the `href` prop and
the `onClick()` prop.

Providing the underlying mouse event to the router
lets it decide when and when not to override the browser's default
behaviour. Providing the `href` prop lets the browser open a link
in a new tab (when applicable) or copy the link.